did some refactoring
